Update yogurt-cake.md w/ sugar
I forgot the sugar, both in this file and in the actual cake I made today. Because I thought the blogger mistakenly added sugar to the wet ingredient list, and I logically intended to add it to the dry. But when I followed the disparate instructions (which is why I write out my own recipes in a way that works for me), sugar of course wasn't listed there, so it didn't get added.